+++ modules/basegl/processors/volumeprocessing/volumegradientmagnitude.cpp
@@ -61,7 +61,7 @@ void VolumeGradientMagnitude::postProcess() {
 
 void VolumeGradientMagnitude::afterInportChanged() {
     if (inport_.hasData()){
-        int channels = inport_.getData()->getDataFormat()->getComponents();
+        int channels = static_cast<int>(inport_.getData()->getDataFormat()->getComponents());
 
         if(channels == static_cast<int>(channel_.size()))
             return;
